+++ examples/tokenvm/orderbook/dependencies.go
@@ -1,0 +1,9 @@
+package orderbook
+
+import (
+	"github.com/ava-labs/avalanchego/utils/logging"
+)
+
+type Controller interface {
+	Logger() logging.Logger
+}
