+++ examples/tokenvm/rpc/consts.go
@@ -1,0 +1,3 @@
+package rpc
+
+const ordersToSend = 128
