+++ examples/tokenvm/rpc/errors.go
@@ -1,0 +1,8 @@
+package rpc
+
+import "errors"
+
+var (
+	ErrTxNotFound    = errors.New("tx not found")
+	ErrAssetNotFound = errors.New("asset not found")
+)
